Extract light channel view helpers
